Fix (RGeo): simple mercator factory has weird behaviour
Failures:

  1) GeoArea#valid? hourglass_polygon is expected to be falsey
     Failure/Error: it { expect(geo_area.valid?).to be_falsey }

       expected: falsey value
            got: true

  2) GeoArea#valid? invalid_right_hand_rule_polygon
     Failure/Error: RGeo::GeoJSON.decode(value.to_json, geo_factory: RGeo::Geographic.simple_mercator_factory)

     NoMethodError:
       undefined method `is_simple?' for nil:NilClass
       Did you mean?  is_haml?

It looks like it only occurs on macOS. This doesn't appear to be an
issue with the rgeo gem but rather the underlying geos library on mac.

There appears to be a difference in the way that geos 3.8.x (3.8.1 is
the latest version in Homebrew circa 4/13/2020) and earlier versions
of the geos library handle the is_simple calculation that determines
Polygon validity. This affects any RGeo factory (most of them but not
RGeo::Geographic.spherical_factory or RGeo::Cartesian.simple_factory)
backed by the CAPI.